update margin to cfd trading
