+++ src/pages/trade-types/margin/_why-trade-cfd.js
@@ -10,45 +10,45 @@ import InstantAccess from 'images/svg/trade-types/instant-access.svg'
 import SyntheticIndices from 'images/svg/trade-types/synthetic-indices.svg'
 import { LinkButton } from 'components/form'
 
-const WhyTradeMargin = () => {
+const WhyTradeCFD = () => {
     return (
         <>
             <SectionContainer background="rgba(242, 243, 244, 0.3)" padding="4rem 0 4rem">
                 <SmallContainer direction="column" ai="flex-start">
                     <Header as="h3" type="section-title" mb="4rem">
-                        {localize('Why trade on margin with Deriv')}
+                        {localize('Why trade on CFDs with Deriv')}
                     </Header>
                     <Grid>
                         <WhyTradeItem>
                             <img src={HighLeverge} alt="high leverage" />
                             <Text weight="bold" mb="0.8rem" mt="1.6rem">
-                                {localize('High leverage, low spreads')}
+                                {localize('High leverage, tight spreads')}
                             </Text>
                             <Text mb="4rem">
                                 {localize(
-                                    'Take advantage of high leverage and low spreads on Deriv MT5 (DMT5).',
+                                    'Take advantage of high leverage and tight spreads on Deriv’s CFD trading platforms.',
                                 )}
                             </Text>
                         </WhyTradeItem>
                         <WhyTradeItem>
                             <img src={SyntheticIndices} alt="synthetic indices" />
                             <Text weight="bold" mb="0.8rem" mt="1.6rem">
-                                {localize('All favourite markets available')}
+                                {localize('All favourite markets')}
                             </Text>
                             <Text mb="4rem">
                                 {localize(
-                                    'Trade on all popular markets plus our proprietary synthetic indices that are available 24/7.',
+                                    'Trade on financial markets plus our proprietary synthetic indices that are available 24/7.',
                                 )}
                             </Text>
                         </WhyTradeItem>
                         <WhyTradeItem>
                             <img src={MaximizePotentialProfit} alt="maximize potential profit" />
                             <Text weight="bold" mb="0.8rem" mt="1.6rem">
-                                {localize('Go long and short')}
+                                {localize('Go long or short')}
                             </Text>
                             <Text mb="4rem">
                                 {localize(
-                                    'Open long and short positions, depending on your preferred trading strategy.',
+                                    'Open long or short positions based on your predictions of the market price movements.',
                                 )}
                             </Text>
                         </WhyTradeItem>
@@ -72,7 +72,7 @@ const WhyTradeMargin = () => {
                         </WhyTradeItem>
                     </Grid>
                     <Text align="left" width="100%" weight="bold">
-                        {localize("Don't have a Deriv.com account yet?")}
+                        {localize("Don't have a Deriv account yet?")}
                     </Text>
                     <LinkButton mt="1.6rem" to="/signup/" secondary="true">
                         {localize('Create free demo account')}
@@ -83,4 +83,4 @@ const WhyTradeMargin = () => {
     )
 }
 
-export default WhyTradeMargin
+export default WhyTradeCFD
